+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pedroandrad1/tokens
 
+## 2.1.2
+
+### Patch Changes
+
+- Separados Toast Component e Toast Provider
+
 ## 2.1.1
 
 ### Patch Changes